Make time reflect shorter deathbans, since they may not always be an increment of 1 hour.

+++ src/main/java/net/frozenorb/foxtrot/server/ServerHandler.java
@@ -346,7 +346,7 @@ public class ServerHandler {
         }
 
         // The default max.
-        long max = Deathban.DEFAULT.inSeconds();
+        int max = Deathban.DEFAULT.inSeconds();
 
         if (player != null) {
             for (Deathban deathban : Deathban.values()) {
